Rag-Techniques/basic-rag/requirements.txt
# Core LangChain packages (exact pins for reproducibility)
langchain==0.2.0
langchain-community==0.0.27

# langchain-core documents support
langchain-core==0.1.34

# OpenAI integration
openai==0.27.0
langchain-openai~=0.2

# HuggingFace embeddings (standalone package replacing deprecated class)
langchain-huggingface~=0.1

# FAISS for vector indexing (MMR retriever, local store)
faiss-cpu~=1.7

# PDF parsing backend for PyPDFLoader
pypdf~=3.11

# Environment variable management
python-dotenv~=1.0
